Add saving location to csv file.

+++ tools/uploader/lib/model/time_computing.dart
@@ -3,7 +3,7 @@ import 'position.dart';
 class TimeComputing {
   const TimeComputing();
 
-  Future<(double, double, Duration)?> getTravelDuration(
+  Future<(int, int, Duration)?> getTravelDuration(
     Position from,
     Position to,
   ) async {
